+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.2.0] - 2024-04-03
+### Added
+- use_undetected=True uses undetected-chromedriver
+- mobile_emulation=True opens in a mobile screen
+
 ## [0.1.0] - 2024-02-12
 ### Added
 - Initial setup of repository (refactored from a private repo)